refactor: replace module `types` with module `_types`
